clyde-python-backend/requirements.txt
absl-py==0.13.0
argcomplete==1.10.0
astunparse==1.6.3
beautifulsoup4==4.8.0
boto3==1.17.96
botocore==1.20.96
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
cycler==0.10.0
dill==0.3.4
docx2txt==0.8
EbookLib==0.17.1
expertai-nlapi==2.3.1
extract-msg==0.23.1
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.4.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
html2text==2020.1.16
idna==2.10
IMAPClient==2.1.0
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
nlp==0.2.0
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pdfminer.six==20181108
Pillow==8.2.0
protobuf==3.17.3
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-pptx==0.6.18
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
sacremoses==0.0.45
sentencepiece==0.1.95
six==1.15.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
stop-words==2018.7.23
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
textract==1.6.3
tokenizers==0.8.0rc4
tqdm==4.61.1
transformers==3.0.0
typing-extensions==3.7.4.3
tzlocal==1.5.1
urllib3==1.26.5
Werkzeug==2.0.1
wordcloud==1.8.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.4.3
youtube-transcript-api==0.4.1
zipp==3.4.1
